+++ tests/unit/Fixtures/DummyMutableCollection.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+
+namespace Headsnet\Collections\Test\Fixtures;
+
+use Headsnet\Collections\AbstractMutableCollection;
+
+/**
+ * @extends AbstractMutableCollection<DummyCollectionItem>
+ */
+final class DummyMutableCollection extends AbstractMutableCollection
+{
+    public function getItemClassName(): string
+    {
+        return DummyCollectionItem::class;
+    }
+}